Code corrections
