+++ l10n_ch_fds_upload_sepa/wizards/__init__.py
@@ -1,4 +1,3 @@
 # -*- coding: utf-8 -*-
 
 from . import payment_order_upload_sepa_wizard
-from . import fds_postfinance_file_pain002
